# opt: refactor streaming decoder to fully use buffer