--- a/internal/decoder/stream.go
+++ b/internal/decoder/stream.go
@@ -17,14 +17,14 @@
 package decoder
 
 import (
-	"bytes"
-	"io"
-	"sync"
+    `bytes`
+    `io`
+    `sync`
 
-	"github.com/bytedance/sonic/internal/native"
-	"github.com/bytedance/sonic/internal/native/types"
-	"github.com/bytedance/sonic/internal/rt"
-	"github.com/bytedance/sonic/option"
+    `github.com/bytedance/sonic/internal/native`
+    `github.com/bytedance/sonic/internal/native/types`
+    `github.com/bytedance/sonic/internal/rt`
+    `github.com/bytedance/sonic/option`
 )
 
 var (
@@ -81,7 +81,7 @@ func (self *StreamDecoder) Decode(val interface{}) (err error) {
             } else {
                 // println("no more")
                 err = SyntaxError{e, self.s, types.ParsingError(-s), ""}
-                self.err = err
+                self.setErr(err)
                 return
             }
         } else {
@@ -94,18 +94,20 @@ func (self *StreamDecoder) Decode(val interface{}) (err error) {
         self.Decoder.Reset(string(self.buf[s:e]))
         err = self.Decoder.Decode(val)
         if err != nil {
-            self.err = err
+            self.setErr(err)
             return 
         }
 
         self.scanp = e
         _, empty := self.scan()
         if empty {
+            // println("recycle")
             // no remain valid bytes, thus we just recycle buffer
             mem := self.buf
             self.buf = nil
             bufPool.Put(mem[:0])
         } else {
+            // println("keep")
             // remain undecoded bytes, move them onto head
             n := copy(self.buf, self.buf[self.scanp:])
             self.buf = self.buf[:n]
@@ -166,12 +168,18 @@ func (self *StreamDecoder) readMore() bool {
 
         // buffer has been scanned, now report any error
         if err != nil  {
-            self.err = err
+            self.setErr(err)
             return false
         }
     }
 }
 
+func (self *StreamDecoder) setErr(err error) {
+    self.err = err
+    mem := self.buf[:0]
+    self.buf = nil
+    bufPool.Put(mem)
+}
 
 func (self *StreamDecoder) peek() (byte, error) {
     var err error
@@ -182,7 +190,7 @@ func (self *StreamDecoder) peek() (byte, error) {
         }
         // buffer has been scanned, now report any error
         if err != nil {
-            self.err = err
+            self.setErr(err)
             return 0, err
         }
         err = self.refill()
@@ -236,8 +244,8 @@ func realloc(buf *[]byte) bool {
     if c - l <= c >> minLeftBufferShift {
         // println("realloc!")
         e := l+(l>>minLeftBufferShift)
-        if e < option.DefaultDecoderBufferSize {
-            e = option.DefaultDecoderBufferSize
+        if e <= c {
+            e = c*2
         }
         tmp := make([]byte, l, e)
         copy(tmp, *buf)