# Optimization
- refactor implementation of `StreamDecoder.Decode()`
- reduce memory copy by limiting start and end of slice passed to `skip()` 
- never move head of decoder buffer, always move tail
- check if buffer remains valid bytes after each`Decode()` or `More()`, if not, recycle buffer
# Benchmark
```
name                          old time/op    new time/op    delta
DecodeStream_Sonic/halt-16       698µs ± 7%     732µs ± 3%   +4.84%  (p=0.002 n=10+10)
DecodeStream_Sonic/single-16     273µs ± 7%     191µs ± 4%  -30.03%  (p=0.000 n=9+10)
DecodeStream_Sonic/4x-16        1.49ms ± 2%    0.59ms ± 7%  -59.98%  (p=0.000 n=10+10)
DecodeStream_Sonic/double-16     679µs ± 3%     237µs ± 5%  -65.00%  (p=0.000 n=10+9)
DecodeStream_Sonic/small-16     27.0µs ± 8%     1.6µs ± 3%  -94.17%  (p=0.000 n=10+9)

name                          old alloc/op   new alloc/op   delta
DecodeStream_Sonic/halt-16       265kB ± 0%     142kB ± 1%  -46.38%  (p=0.000 n=10+10)
DecodeStream_Sonic/double-16    1.20MB ± 0%    0.28MB ± 0%  -76.65%  (p=0.000 n=10+10)
DecodeStream_Sonic/4x-16        2.79MB ± 0%    0.56MB ± 0%  -79.87%  (p=0.000 n=10+10)
DecodeStream_Sonic/single-16    1.32MB ± 0%    0.14MB ± 0%  -89.36%  (p=0.000 n=10+9)
DecodeStream_Sonic/small-16      269kB ± 0%       1kB ± 0%  -99.45%  (p=0.000 n=10+10)

name                          old allocs/op  new allocs/op  delta
DecodeStream_Sonic/small-16       21.0 ± 0%      20.0 ± 0%   -4.76%  (p=0.000 n=10+10)
DecodeStream_Sonic/halt-16        12.0 ± 0%       9.0 ± 0%  -25.00%  (p=0.000 n=10+10)
DecodeStream_Sonic/double-16      18.0 ± 0%       9.0 ± 0%  -50.00%  (p=0.000 n=10+10)
DecodeStream_Sonic/4x-16          29.0 ± 0%      14.0 ± 0%  -51.72%  (p=0.000 n=10+10)
DecodeStream_Sonic/single-16      11.0 ± 0%       5.0 ± 0%  -54.55%  (p=0.000 n=10+10)
```